relocate code
